Sprint 5

+++ LibraryWebParent/LibraryBackend/LibraryReader/src/main/java/gb/library/reader/controllers/ReservedBookController.java
@@ -4,8 +4,11 @@ import gb.library.common.entities.ReservedBook;
 import gb.library.reader.converters.ReservedBookConverter;
 import gb.library.reader.dtos.ReservedBookReaderDto;
 import gb.library.reader.services.ReservedBooksService;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -17,18 +20,22 @@ public class ReservedBookController {
     private final ReservedBookConverter converter;
 
     @GetMapping("/{userId}")
-    public Page<ReservedBookReaderDto> findAll(@RequestParam(name = "page") int pageIndex,
+    public Page<ReservedBookReaderDto> findAll(@RequestParam(name = "page", defaultValue = "1", required = false) int pageIndex,
                                                @PathVariable int userId) {
         return reservedBooksService.getAllPageable(userId, pageIndex).map(converter::entityToDto);
     }
 
-    @DeleteMapping("/delete/{id}")
-    public void deleteFromReserved(@PathVariable int id) {
+    @DeleteMapping("/{id}")
+    public ResponseEntity<?> deleteFromReserved(@PathVariable int id) {
+        if (id < 1)
+            return ResponseEntity.badRequest().build();
+
         reservedBooksService.delete(id);
+        return ResponseEntity.noContent().build();
     }
 
-    @PostMapping("/create")
-    public ReservedBookReaderDto reserveBook(@RequestBody ReservedBookReaderDto reservedBookDto) {
+    @PostMapping
+    public ReservedBookReaderDto reserveBook(@RequestBody @Valid ReservedBookReaderDto reservedBookDto) {
         ReservedBook book = reservedBooksService.create(converter.DtoToEntity(reservedBookDto));
         return converter.entityToDto(book);
     }
